ai_model/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.6,<1.0.0
pydantic>=2.9.2,<3.0.0
numpy>=2.1.0,<3.0.0
scikit-learn==1.5.2
joblib>=1.4.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0